# contracts/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Bar Buddy API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/api/v1/cocktails" : {
      "get" : {
        "operationId" : "listCocktails",
        "parameters" : [ {
          "in" : "query",
          "name" : "search",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "primarySpiritId",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "availability",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "favoritesOnly",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CocktailSummary"
                  }
                }
              }
            },
            "description" : "Catalog result"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Invalid catalog input"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication is required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Catalog item not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "catalog-browse-controller" ]
      }
    },
    "/api/v1/cocktails/{id}" : {
      "get" : {
        "operationId" : "getCocktail",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CocktailDetail"
                }
              }
            },
            "description" : "Catalog result"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Invalid catalog input"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication is required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Catalog item not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "catalog-browse-controller" ]
      }
    },
    "/api/v1/cocktails/{id}/preference" : {
      "put" : {
        "operationId" : "updateCocktailPreference",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateCocktailPreference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CocktailPreferenceResponse"
                }
              }
            },
            "description" : "Cocktail preference updated"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Invalid catalog input"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication is required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Catalog item not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "catalog-browse-controller" ]
      }
    },
    "/api/v1/ingredients" : {
      "get" : {
        "operationId" : "listIngredients",
        "parameters" : [ {
          "in" : "query",
          "name" : "search",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "category",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/IngredientSummary"
                  }
                }
              }
            },
            "description" : "Catalog result"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Invalid catalog input"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication is required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Catalog item not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "catalog-browse-controller" ]
      }
    },
    "/api/v1/ingredients/{id}" : {
      "get" : {
        "operationId" : "getIngredient",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IngredientDetail"
                }
              }
            },
            "description" : "Catalog result"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Invalid catalog input"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication is required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Catalog item not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "catalog-browse-controller" ]
      }
    },
    "/api/v1/inventory" : {
      "get" : {
        "operationId" : "listInventory",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/InventoryResponse"
                  }
                }
              }
            },
            "description" : "Your inventory"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Invalid inventory input"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Inventory item not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "inventory-controller" ]
      },
      "post" : {
        "operationId" : "createInventory",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateInventory"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InventoryResponse"
                }
              }
            },
            "description" : "Inventory item created"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Invalid inventory input"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Inventory item not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "inventory-controller" ]
      }
    },
    "/api/v1/inventory/{id}" : {
      "delete" : {
        "operationId" : "deleteInventory",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Inventory item removed"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Invalid inventory input"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Inventory item not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "inventory-controller" ]
      },
      "patch" : {
        "operationId" : "updateInventory",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateInventory"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InventoryResponse"
                }
              }
            },
            "description" : "Inventory item updated"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Invalid inventory input"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Inventory item not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "inventory-controller" ]
      }
    },
    "/api/v1/me" : {
      "get" : {
        "operationId" : "getCurrentUser",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "The current Bar Buddy user"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemResponse"
                }
              }
            },
            "description" : "Authentication is required"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Resolve the current Bar Buddy user",
        "tags" : [ "Users" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "ApiProblemResponse" : {
        "type" : "object",
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string",
            "format" : "uri"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "format" : "uri"
          }
        },
        "required" : [ "detail", "instance", "status", "title", "type" ]
      },
      "AvailabilityResult" : {
        "type" : "object",
        "properties" : {
          "canMake" : {
            "type" : "boolean"
          },
          "missingCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "missingIngredients" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IngredientSummary"
            }
          }
        }
      },
      "CocktailDetail" : {
        "type" : "object",
        "properties" : {
          "availability" : {
            "$ref" : "#/components/schemas/AvailabilityResult"
          },
          "favorite" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "primarySpirit" : {
            "$ref" : "#/components/schemas/IngredientSummary"
          },
          "recipe" : {
            "$ref" : "#/components/schemas/RecipeDetail"
          },
          "slug" : {
            "type" : "string"
          }
        }
      },
      "CocktailPreferenceResponse" : {
        "type" : "object",
        "properties" : {
          "cocktailId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "favorite" : {
            "type" : "boolean"
          }
        }
      },
      "CocktailSummary" : {
        "type" : "object",
        "properties" : {
          "availability" : {
            "$ref" : "#/components/schemas/AvailabilityResult"
          },
          "favorite" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "primarySpirit" : {
            "$ref" : "#/components/schemas/IngredientSummary"
          },
          "slug" : {
            "type" : "string"
          }
        }
      },
      "CreateInventory" : {
        "type" : "object",
        "properties" : {
          "bottleLabel" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          },
          "ingredientId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "Have", "Out" ]
          }
        },
        "required" : [ "ingredientId", "status" ]
      },
      "DisplayMeasurement" : {
        "type" : "object",
        "properties" : {
          "maximumQuantity" : {
            "type" : "number"
          },
          "modifier" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "number"
          },
          "unit" : {
            "type" : "string"
          }
        }
      },
      "IngredientDetail" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "relatedCocktails" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CocktailSummary"
            }
          },
          "usageCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "IngredientSummary" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "matchedAlias" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "InventoryResponse" : {
        "type" : "object",
        "properties" : {
          "bottleLabel" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ingredient" : {
            "$ref" : "#/components/schemas/IngredientSummary"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "Have", "Out" ]
          }
        }
      },
      "MeResponse" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "createdAt", "id" ]
      },
      "RecipeDetail" : {
        "type" : "object",
        "properties" : {
          "garnish" : {
            "type" : "string"
          },
          "glassware" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ingredients" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RecipeLine"
            }
          },
          "instructions" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "RecipeLine" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "ingredient" : {
            "$ref" : "#/components/schemas/IngredientSummary"
          },
          "metric" : {
            "$ref" : "#/components/schemas/DisplayMeasurement"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          },
          "preparation" : {
            "type" : "string"
          },
          "requirement" : {
            "type" : "string"
          },
          "us" : {
            "$ref" : "#/components/schemas/DisplayMeasurement"
          }
        }
      },
      "UpdateCocktailPreference" : {
        "type" : "object",
        "properties" : {
          "favorite" : {
            "type" : "boolean"
          }
        },
        "required" : [ "favorite" ]
      },
      "UpdateInventory" : {
        "type" : "object",
        "properties" : {
          "bottleLabel" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          },
          "status" : {
            "type" : "string",
            "enum" : [ "Have", "Out" ]
          }
        },
        "required" : [ "status" ]
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}